server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openenv-core[core]>=0.2.2
pydantic>=2.0.0
httpx>=0.24.0
python-multipart>=0.0.6
